Module-6/Demo-7/enterprise-mcp-weather/requirements.txt
mcp
fastmcp
requests
python-dotenv
tenacity
streamlit
ollama